[578510] Fix can_tag permission's absence and many other neglected migrations.
I couldn't agree with myself whether to use can_tag or tag_question as the permission's codename. Fixed that.

Also, we've been neglecting from the beginning to add entries to django_content_type and auth_permission when we add new models. Since we don't run syncdb, this isn't taken care of automatically. Went back and added the right things to old migrations.