Make SUMO's WikiParser a subclass of wikimarkup's Parser.
This is in preparation for showfor implementation, which uses a custom tag handler, which receives a copy of the parser when called. With the previous composition-based wrapping, the uncustomized parser would be passed to the tag handler, and any markup inside the tag would be parsed in an un-SUMO-like fashion.