[586478] New home page
Create a new home page at /home/ that pulls in editable content from
the wiki. Involves some tweaks to the wiki templates. Also add the rest
of the KB categories.

* Hard coded top side can be localized with gettext
* Not tested for IE
* Migration to create initial templates
* Inherits from wiki base template

As the new design propagates to the rest of the site, we need to make
the common/layout templates the base of both wiki and home page.