Anyone can create new revisions. [bug 610421]
I took the absolute minimun change to make this work, leaving the
Document.allows_revision_by method in place, in case we do end up with
different permissions for, e.g., templates.

If that code rots for a while and it looks like we'll never do that,
we can take a bigger patch when there's more time and remove it
completely.