[bug 627359] Make get_object_fallback() follow redirects when failing back to the default locale.
This means that, for example, if you have Document D in Deutsch which links to [[A]], A is a redirect to B, and B has a Deutsch translation Bd, rendering D will give you a link straight to Bd. That's about as simple as it gets, I'm afraid, but it's really handy when English articles get renamed and tranlations linked straight to the English.

* A large chunk of DocumentVisit._visit_counts() got turned into Document.from_url(), where redirect_document() can use it too.
* Add Document.redirect_document().
* Move all the get_object_fallback_* tests into their own testcase. The super-long function names were driving me crazy.
* WebTrends update cron now limits its interests to Documents with an approved revision.