Edit a localized document based on a previous revision [bug 624416]
When you edit a localized document 'based on' a previous revision of the
*localized* document (i.e. not of the English doc) we were setting the
'based_on' value to a revision of the localized doc, not the English doc.

Now we set the 'based_on' revision to the correct English document but
still prefill the form with the revision of the localized document you're
editing.

Whew.

STR:

* Go to any localized article's history.
* Pick any revision and 'edit based on this revision'.
* Attempt to submit.

Without this patch, you'll see an error. With it, everything should work.

Also worth noting that the new test fails on master but passes here.