Bug 665714 - Kuma: "Save and keep editing", without leaving edit mode
Introduces save-and-edit button that retargets the editing form to
submit to a hidden iframe. The normal save button submits to the top
level page. If a login-required or 403 forbidden is encountered, the
page busts out of the iframe to the top level.

Additionally, edited content is preserved in sessionStorage when
save-and-edit is clicked. The content gets discarded on a successful
save, restored to the textarea on a return to the page after logging
back in.