Enforce uniqueness of slugs for long demo titles, and clean up corresponding tests.
Slug uniqueness is handled by looking for another demo with the same
slug, and then appending numbers (00-99) to the end of the slug until
a unique slug occurs or we run out of numbers (which hopefully doesn't
happen).

This ended up breaking some tests, because they all used a
"hello-world" slug and assumed they could just create it over and over
and have it come out the same. I've modified save_valid_submission()
to do smarter slugification and invoke this, and tweaked one test that
had an expected slug hard-coded into it; it now uses whatever slug the
Submission object actually ends up getting.