Fix bug 671773: Feed of recent revisions
This provides a feed of the 50 latest Revisions made in the
wiki. Individual Revisions show up as follows:

* For the first Revision of a new Document, link to the Document with
  a 'document created' description.

* For Revisions of an existing Document, link to the compare view of
  this + previous Revision, with the edit comment as description.

Potential issue: there's no good way to share state amongst the feed
methods for an individual item, meaning we repeat the logic to figure
out if there's a previous Revision of the same Document. This makes
for a possibly-ugly number of DB queries, but I couldn't find a simple
way to get around that.