Files #1: file infrastructure.
This is the most basic low-level infrastructure for file
handling. It's not feature-complete and it's not particularly pretty,
but it's enough to build on and to support the other chunks of file
code that are or will be in the pipeline.

With this code, we have the concept of a file attachment and revisions
of that file, and a URL and a very naive view for serving the
file. The latter will inevitably be replaced by more detailed
infrastructure, and is only here to provide a bit of scaffolding to
allow the basics of file handling to be inspected manually while that
infrastructure is being built.