This adds two models -- Attachment and AttachmentRevision -- which represent a file and its revisions. Attachments have a concept of "current revision" just like Documents do, and any arbitrary revision can be made the current one.
Attachments store a MindTouch ID, but not solely for bookkeeping
purposes. Because MindTouch's file URLs use the resource ID as a
unique identifier, we have to keep that in order to redirect MindTouch
URLs correctly. Their revisions store a MindTouch ID solely for future
migration bookkeeping; that field can be deleted later.

A ludicrously simple and possibly broken file-serving view is present,
and should be replaced with real views (coming in a future set of
commits).

Documents have a new property -- 'attachments' -- which returns a
queryset of Attachments for the current revision of the Document. At
the moment this is done by scanning the HTML for URLs matching the
MindTouch file URL pattern. This will, once kuma's own real file URLs
get settled, also look for those.