Bug 761826: Handle legacy MindTouch file URLs.
This is simply a URL pattern and corresponding view which responds to
the MindTouch file URL scheme, looks up the correspondong kuma
Attachment, and issues a redirect.

It is relatively future-proof for the case where kuma's own file URL
patterns change (very likely), and is testable independent of having a
fully-migrated set of files (very handy). Test cases are actual
MindTouch URLs, with corresponding dummy Attachments created so kuma
has something to redirect to.