MindTouch-to-kuma file migration script.
Does what it says on the tin: adds '--files' to the list of
command-line flags for the migration script, and does everything in
its power to migrate files and their revisions into kuma's
file-handling infrastructure.

Currently, on a somewhat dated snapshot of the DB and files, it misses
three revisions.