Work around duplicate migrated deki users.
This is at best a temporary hack to get migration working again; it
simply changes out the get() call for filter(), and automatically
picks the first user returned. This may result in a need to fix
objects associated with a second account, but since currently the
migration blows up when attempting to call this method, hopefully it's
not been possible to create any such objects.