Bare-bones feed of file edits.
This is almost entirely lifted from the existing feed of revisions of
documents, tweaked a bit to match the way the file models are laid
out. To make it easier to determine if a given file revision is a new
file, a get_previous() method is added to AttachmentRevision (and
probably should've been there anyway).