Remove ability to delete revisions, and add reversion.
The delete-revision view still exists after this, but is no longer
wired up anywhere, and the document history template no longer offers
the ability to use it.

A new method on Document and a new view for reverting are added,
taking a revision ID and the user who's performing the revert, and
creating a new Revision based on that revision ID with a "revert"
comment.