Added link to Bugzilla from login page.
A few people have been unable to log in with Persona lately. Most often,
this is because they do not have an email address associated with their
account and additionally cannot remember their former account password.

The fix is simple -- manually set their email address -- but they need
to get in touch with us to make that possible. This link should help
them do that.