Fix bug 785212: Method for getting all children of a page
This is a somewhat naive recursive implementation that descends
through children, grandchildren, etc. and builds up a data structure
that clearly shows the relationships and sub-trees and so forth.

It is naive in the sense that A) it does not attempt to minimize the
number of DB queries involved in building the tree, and B) currently
has no way at all of detecting a circular dependency.