Add a bit of special-case logic in tree-moving.
The slug manipulation for tree moves is implemented using basic string
replace operations, but this fails on the "special" (albeit likely
common) case of simply wanting to stick a prefix onto the beginning of
the slug. This could be implemented by resorting to regular-expression
replacement that's aware of the position to insert at, but it's far
simpler to just treat this as a special case.