Preserve tags and review tags when moving.
Tags were already being kept because they're a fairly normal field
(though now a test proves this). Review tags are a bit more complex,
and now are explicitly stored temporarily during page moving so that
they can be added back to the new revision that comes out the other
side.