Fix bug 792955: Build file upload filter
This adds a mime-type check to the process of uploading an attachment
to the wiki, and will disallow files outside of a specific whitelist
of permitted types.

The whitelist is specified through constance, with a simple default of
just the common image formats, but as a constance key it'll be
editable through the admin without needing to tweak code.