Page move frontend.
This is not pure "frontend" in the sense that it includes more than
just HTML/CSS/JS. It also provides the form and view handling for
page-move, plus some necessary infrastructure to make them work.

In a reversal of a previous version of this code, attempting to edit
the slug of a document with children will not immediately kick into
the page-move interface; instead, it will notify the user, through a
form validation message, that a move is required.