Page move frontend.
This is not pure "frontend" in the sense that it includes more than
just HTML/CSS/JS. It also provides the form and view handling for
page-move, plus some necessary infrastructure to make them work.

In a reversal of a previous version of this code, attempting to edit
the slug of a document with children will not immediately kick into
the page-move interface; instead, it will notify the user, through a
form validation message, that a move is required.

Note also that the slug-editing tests heavily rely on behaviors which
are obsoleted by page move. I've made multiple attempts at revamping
them to not rely on the ability to edit a document with children, but
in the name of getting this working I finally just took a blunt-object
approach to those tests. So they're still there, but not currently
being run until we can actually rewrite them to not run straight into
the page-move validation process.