Rework slug generation and propagation in page move.
This is far simpler than the previous iteration, and so is also a lot
more robust; page moving can now handle pages which don't exist within
a strict URL hierarchy, because it will forcibly put them into such a
hierarchy. The flipside is that it is now an ironclad assumption that
a moved hierarchy of pages will always have its parent/child
relationships reflected in the resulting URLs.

As a side note, some of the unit tests now look a bit strange because
-- even though they're showing the correct behavior -- the old slugs
that were being used don't always clearly fit the above assumption. A
possible TODO is to clean those up.