Bug 900961: Provide a way to (sometimes) repair bad breadcrumbs.
This doesn't prevent the problem from recurring; doing that will
probably require digging down into the page-move code and figuring out
what exactly went wrong that broke the breadcrumbs in the first place.

This also will not fix every broken document; some are too bad for
this sort of fix to work on, and will have to be repaired manually.

For ease of UI, a view is provided, accessible to superusers, which
can be linked from a document detail page to perform the repair.