Fix bug 903723 and bug 904494: Revision dashboard cleanup.
This fixes a couple of unfortunate typos that got through all of: me,
the unit tests, and the code review on github. The primary issue is
that the unit tests for the revision dashboard were not truly
inspecting the results; this commit fixes the view issues and adds a
bunch of revisions in a fixture, and then has the tests -- both for
the general listing and for various filterings -- make sure that the
correct number of results come back, so that we have a way to catch
regressions going forward.