Migrate some mdn-screen rules to Stylus
Some infrastructure work is included here. Long story short, the
css_migration flag is added to enable or disable the effects of the CSS
migration. This flag should not be enabled for all users on production,
but it might make sense to enable it for select users so that they have
the opportunity to find and report on unintentional side-effects of the
migration.

When the flag is disabled, users should not see any visual changes. The
site should appear just as it did before this pull request landed.

When the flag is enabled, users should /still/ not see any visual
changes (the goal of the migration is to move rules to Stylus without
changing the visual design) but they might see some unintentional
side-effects of this migration. For example, if I incorrectly deleted a
rule that is actually still needed, they would see some issues.

A new stylesheet named mdn-screen-old.css is added to achieve this.
Rules that are thought to be obsolete are moved here, and the stylesheet
is only loaded for those people who are /not/ testing the migration. The
stylesheet is not loaded for people who /are/ testing the migration.

The cost is one additional HTTP request. I can live with that for now.