Issue #1755: Remove pointer from compat tab hover
This is one part of #1755. Many other changes were made on the
production Template:CustomCSS that also address this issue. The issue
has not been completely resolved (there are small areas near the edges
of each tab that cannot be clicked) but this is about as far as I can go
for now. If we ever clean up Template:CustomCSS, the rest will be easier
to fix.