Update, remove old design styles in mdn-screen.css
Rules that are still needed are moved to Stylus and updated to work
correctly alongside styles from the new design. Along the way, page
elements and even entire pages are updated to adhere to conventions of
the new design. Spacing is adjusted, colors are updated, responsive
layouts are added, and more.

The design of the Demo Studio is not updated (and Demo Studio rules are
not moved to Stylus) because a project is already planned to completely
redesign these pages. Significant user-facing improvements are made to
the learning pages, but some learning page rules are copied directly
from mdn-screen.css and left untouched. The quality of the underlying
implementation matters very little, in my opinion, as the next major
user-facing change to these pages will likely be removal.