Migrate mdn-screen.css to Stylus
Some groundwork is included here. To summarize, the css_migration flag
is added to enable the effects of the migration. When the flag is
enabled, a modified version of mdn-screen.css (with obsolete and
migrated rules removed) is attached to pages. When the flag is disabled,
the normal mdn-screen.css stylesheet is attached instead.

The flag should be disabled for almost all users, but it might make
sense to enable the flag for beta testers so that they can find bugs in
the migration and report on them.

The cost is one additional request. Worth it for an implementation that
is simple and understandable, in my opinion.