Added Foreman Procfile for easier development running.
Foreman is the tool of choice for manageing processes on Heroku, so it is “battle tested”.
It starts all or one processes as defined in the Procfile, e.g. `foreman start` starts all. `foreman start web` starts webserver etc.

This needed some other minor changes that were preventing a few things:

- made the /etc/profile.d/vagrantruby.sh file exectuable to automatically append $PATH with the path to the Ruby environment (/opt/ruby/bin)
- fixed the /etc/sudoers to not prevent running ruby installed executables (in /opt/ruby/bin) when using sudo **this may need two runs of puppet to fix it**
- fixed trailing whitespaces in puppet files
- allowed using the puppet gem provider