Do not override "clear" property with custom mixin
Until now, we used a mixin named clear() to apply clear-fixes to
elements.

When Stylus encounters a property that shares a name with a mixin, it
assumes the property is a transparent call to the mixin [1]. Thus,
whenever we used the real CSS property "clear", Stylus assumed we were
calling the mixin of the same name, and we ended up applying clear-fixes
to elements accidentally.

To test this, inspect a Zone landing page title. Without this change,
the title should be incorrectly clear-fixed. With the change, the title
should correctly not be clear-fixed.

[1] http://pages.citebite.com/e2t7l2p4m9dig