Put advanced profile features behind gear submenu
My hope was to A/B test this change with Optimizely, but we are
currently unable to run tests on page elements that only logged-in users
see.

When we made a similar change to article pages (moving the history
feature behind a gear icon) the edit button saw a statistically
significant increase in click rate, so we can be moderately confident
that the same thing will happen here.