Revert cced5a201be5348f1f32cbbdaa61e922cf4b2bdb
Well, it turns out 2f62c0ceee7b2745276082b62c111fd57d3d174e did not
cause the bug after all. I still have no idea why reverting it fixed the
problem on stage, but it seems likely that I arrived at a false positive
in my rush to test and deploy a fix.

And with that, I humbly submit a pull request to revert a revert that
did not cause a problem. //developershame