Fix bug where profile submenu is shown by default
As part of implementing the advanced submenu used on the profile page
(#2028), the .submenu rule was moved from wiki.styl to components.styl
so that it would be available to both wiki and profile pages. When the
submenu system was refactored (#2044), the rule was moved back to
wiki.styl, making it unavailable to profile pages and therefore breaking
profile submenus.

This change moves the rule back to components.styl so that the rule is
available to both.