Adding the promote-mdn embeddable js script
Removed dependancy on jQuery

Removed jQuery dependancy and use of gloabl for settings.

Replacedd massive if block with an early return if browser doesn't support querySelectorAll.

Repalced all under_scored names to camelCased

Added provision to pass a custom classname to be used on outbound links to MDN.

Cleaned up nodelist.forEach a little and added a polyfill for array.forEach

Some stupid bug fixes and improvements in keyword mathcing.

Renamed maxlinks param to maxLinks for consistency.

Fixed typo in a variable name.

Rewriting the array forEach shim code to use a standalone function, rather than extending the array prototype.

Moved the checking for document.querySelectorAll support to the beginning of the code.

Adding provision for the end user to pass extra keywords and links for the script to linkify.

A couple of minor improvements.