Create My interest
I really sympathizing with every one and I'm also sorry for those believed that are thieves. To bring to your attentions that they are just but stupid fools who believed on thorax and capable of killing to survive no mater what the consequences will be. In real life you find people of different life styles. every man focus on how to survive in this wold.
My problem is that I hate real, they know to play with people physiology; they do want to know either your rich or poor, cleaver or fool either your are thin or fat, professor or student, pastor or drug killer or who ever you think your. They manovour your brain as if your his/her child. and that is they way real thieves are stilling your wealth be it cash or solid it will go.
My advice to you is that watch out both internet pretenders and real people you interact with in every towns and cities you've walk in this wold.
Any way have nice time as you think deeper.