Fix bug 972533: Enhancements to admin handling of deleted pages.
This adds a new manager to Document, which does not perform any
filtering based on deleted status, and uses that manager for the
admin; this ensures the admin always shows all Documents currently in
the database.

Building on that, the admin filters now include one to show only
deleted Documents.

The admin actions for Document are also updated:

* A new "purge" action is added, to bulk-purge deleted Documents.

* A new "undelete" action is added, to bulk-undelete deleted
  Documents.

* The built-in Django "delete" action is removed, since we want people
  using the standard deletion UI (and because the admin delete action
  bypasses the model delete() method, which is not what we want for
  deleting Documents).