Fix bug 967017: Re-render moved pages post-move.
This just does a brute-force re-render of every page affected by the
move, as the last part of the page-move task. This will address both
the issue in the bug (which is stale JSON -- that gets rebuilt during
the render) as well as any other stale data (it may also, as a side
effect, fix bug 973605 by rebuilding the list of translations).

Note that this doesn't clean up all stale-JSON issues; post-page-move,
for example, JSON may still be stale for the old URL.