Fix bug 971073: Clear cache on just-deleted document.
The root issue here is not actually whether the document gets
re-rendered; it's that viewing the document causes last-modified info
to go into the cache, so that only a force-refresh afterward clears
it. This adds a cache bust of the document's last-modified info to the
delete() method to ensure we show an uncached "this document has been
deleted" immediately.