Simplify registration page copy
Because we no longer offer the "Forgot Password" form, the only function
of the page is registration. The page title is updated to reflect this
and the "New MDN Members" heading is removed.

Other changes are made, mostly minor. The Bugzilla link is removed -- we
only receive about one report per month, so it seems unnecessary to
burden all users with this notice. People who deeply care about getting
their old accounts back will likely submit a bug with or without this
notice.