Prevent long list of tags from breaking translate page.
Long lists of tags were being output without spaces and if the single line was long enough it would break the page layout.

Fixed at the output source assuming .tags is an object, however if it is a string it might be better for us to add the space in earlier in the processing of the output.