Show signup form errors inline and render non-field form errors as error notifications.
This also makes sure we show the duplicate email error message inline instead of a notification to normalize the error handling between an error with the other email address and one of the Github provided verified email addresses. This is done by demoting the duplicate email form error to a field level error.

Fixes bug #1050008.