Bug 1054558: Notification after successful connect
If a user tries to authenticate with a service that is not already
associated with their account (for example, GitHub) they might be
notified that there is already an MDN profile using the same email
address. When this happens, the user is given the option of
authenticating with a third-party service that /is/ associated with the
account (for example, Persona) to make both services (in this case,
GitHub and Persona) authentication options for that account.

These changes notify the user when this is done successfully.