Bug 1018707 - Empty CustomCSS
Copied CustomCSS into this file so it is minified. Myself and @groovecoder have "subscribed" to changes to CustomCSS and when changes to the file occur they can be moved into a proper place in the code base. We can work through this file slowly integrating its contents into our code base but in the mean time at least it's minified.

When this goes into staging the contents of CustomCSS need to be emptied for testing.

When this goes into production they contents of CustomCSS will need to be removed there as well.