Bug 1070001 - Re-org Stylus - Tags
All tags now use same class ".tags".

There is a modifying class ".tags-small" for the tags in the wiki footer.

There is a .tags_tagit file to extend this with over-rides for the jQuery UI plugin.

Things that are not tags (list of interests on profile page) no longer use the class ".tags".