Fix Bug 1071591 - Download description alignments
To keep headings aligned at the top I removed vertical-align:centre and to keep buttons aligned at the bottom I absolutely position them at the bottom in a space created by adding padding to the bottom of the description.

Also added a fix to keep the "headings" on a separate line on mobile since they look weird on the same line as the image.