Fix Bug 1018596: Previous and Next Buttons
Changed the previousPage and NextPage macro to include the class "neutral" so we can remove the custom styles making these blue. They still need the styles for positioning though. Yay less code!

To test: BEFORE you check this branch out, edit the previousPage and next Page macro to include the "neutral" class on the button. Have a look at a page using the macro and confirm for me that it is okay to have that class on the buttons before the style sheet change is in place.

Then checkout the branch and have a look at the result of the new styles.