Bug 1106059: Disable admin Document deletion.
We already have the delete-from-object-list action disabled; this
simply completes that process by disabling the delete button on
individual Documents, by hard-wiring a lack of delete permission for
every user.

As a result, all Document deletion will be forced through the custom
kuma deletion infrastructure, which ensures we have an auditable trail
of who has deleted a Document and why.