BUG 1070001: Re-Org Stylus - wiki
Fix for no-details. These declarations were changed in a PR between when I created this branch and today, they were probably miss placed in a rebase.

In light of this I'm going to compare our changes to wiki.styl between when I started on this and today to make sure all changes have been carried forward.