Bug 1082031 - Get rid of the wheelballs.
The wheelballs served a good purpose at the time but we've no have all the dependencies locked down in local vendor submodules and can better install them in Vagrant and Travis.

This is needed for this ticket since we want to enable us to install the new compiled components librabbitmq correctly.