Bug 1116337: Clear Django 1.5 lint output.
WARNING: This commit breaks compatibility with Django 1.4.

Many of the warnings from the linter are actually spurious/overzealous
reminders about checking the existence of cleaned_data (bad), where we
actually were checking for the presence of specific elements in
cleaned_data (good), but which django-compat-lint can't handle. As
such they were left as-is, and only the actually-relevant changes
(which are fewer in number) have been made.