Fix bug 1139670: Remove the maintenance page
This is nice to have, and I honestly wish we used it more than we do.
It's much nicer than the standard black and white error message we
usually serve up when the site is down. It worked well, too, even though
IT had an unrelated hosting issue the last time they used it.

That said, we haven't been maintaining it (no pun intended) and we don't
have any maintenance planned. We can always use this code as inspiration
for future maintenance pages, but there's no good reason to let it
bitrot in the repository in the meantime.