Fix bug 1139947: Fix homepage Owl Carousel error
This bug only manifests locally. The "home" bundle is attached to the
homepage asynchronously. The scripts in the bundle still need to be
loaded in order, however.

On production, the scripts *are* loaded in order because the are
concatenated before the whole thing is attached asynchronously.

Locally, each script within the bundle is individually attached
asynchronously, so they're sometimes loaded out of order, causing the
error.