Bug 1139508: Redirect media/redesign with htaccess
We need to redirect from /media/redesign to /media. This htaccess
redirect won't have any affect locally (because we alias /media in
mozilla-kuma-apache.conf), but it might work on production where the
alias isn't active.

Unfortunately, the only way to know for sure is to try it out. If this
doesn't do the trick, we can revert this change and work with WebOps to
set up the redirects in whatever form their configuration requires.