Bug 1116335: Clear test failures in wiki/test_content.
There were several underlying problems here. One -- easy to fix -- had
to do with the fact that not all places where html5lib returned
namespaced values were appropriately dealing with the namespaces.

Another was the use of string comparison on generated versus expected
HTML, which was tripped up by inconsistent ordering of
attributes. These have been replaced with Django's assertHTMLEqual,
which performs a comparison that takes HTML semantics into account.

Another was that DekiscriptMacroFilter was badly broken, but now is
unused in kuma. So it was simply removed, along with the tests for it.

Finally, SectionIDFilter contained processing of elements which was
non-exhaustive, and prone to fail when fed inputs from the uncovered
cases (most notably, when certain ID-generating elements had no
attributes set). SectionIDFilter's logic was modified to ensure it is
now exhaustive.

As a side effect of figuring this out, explanatory comments were added
to SectionIDFilter, trying to show what exactly is going on in there.