Bug 1116335: Modernize and future-proof URLconfs.
This removes all use of the patterns() function, prefixes, string view
names and bare-tuple URL patterns in favor of defining urlpatterns as
a list, importing views into the URLconf to refer to them, and using
url() on all patterns.

Going forward, the preferred structure of any URLconf in kuma should be:

    from django.conf import include, url

    from . import feeds
    from . import views

    urlpatterns = [
        url(r'some_pattern',
            views.some_view,
            name=some_name),
        # ... etc.
    ]

Omitting the import of include() and of a feeds module as appropriate.