bug 957802 - Improve mysql init script
/etc/init.d/mysql greps for the mysql process to determine if mysql is
running.  When running, the process count includes the mysqld process
and the grep for that process.  However, 5% - 10% of the time, the grep
is not in the process list, causing false negatives when detecting the
mysqld process is running.  This new init script ignores the grep
process, making the run detection more reliable.