Bug 1198149 - Separate Jinja2 macros that require context.
This is due to the import behavior of Jinja2: http://jinja.pocoo.org/docs/dev/templates/#import-context-behavior

Macro imports are cached and not passed the template context and hence breaking the csrf token helper. We saw that as a warning in the test output but seem to have missed it in production.